cc-3319 float fixes for product packaging unit

+++ Bundles/ProductPackagingUnit/src/Spryker/Zed/ProductPackagingUnit/Business/Model/ProductPackagingUnit/ProductPackagingUnitAmountSalesUnitValue.php
@@ -57,13 +57,13 @@ class ProductPackagingUnitAmountSalesUnitValue implements ProductPackagingUnitAm
     }
 
     /**
-     * @param int $availabilityValue
+     * @param float $availabilityValue
      * @param float $unitToAvailabilityConversion
      * @param int $unitPrecision
      *
      * @return int
      */
-    protected function calculateNormalizedValue(int $availabilityValue, float $unitToAvailabilityConversion, int $unitPrecision): int
+    protected function calculateNormalizedValue(float $availabilityValue, float $unitToAvailabilityConversion, int $unitPrecision): int
     {
         return (int)round(
             $this->calculateFloatNormalizedValue($availabilityValue, $unitToAvailabilityConversion, $unitPrecision)
@@ -71,13 +71,13 @@ class ProductPackagingUnitAmountSalesUnitValue implements ProductPackagingUnitAm
     }
 
     /**
-     * @param int $availabilityValue
+     * @param float $availabilityValue
      * @param float $unitToAvailabilityConversion
      * @param int $unitPrecision
      *
      * @return float
      */
-    protected function calculateFloatNormalizedValue(int $availabilityValue, float $unitToAvailabilityConversion, int $unitPrecision): float
+    protected function calculateFloatNormalizedValue(float $availabilityValue, float $unitToAvailabilityConversion, int $unitPrecision): float
     {
         return $availabilityValue / $unitToAvailabilityConversion * $unitPrecision;
     }
